requirements.txt
annotated-types==0.6.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.8
cymem==2.0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.7.0/es_core_news_lg-3.7.0-py3-none-any.whl
filelock==3.13.1
Flask==3.0.3
idna==3.6
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
langcodes==3.3.0
lxml==5.2.2
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pdfminer==20191125
pdfminer.six==20231228
phonenumbers==8.13.32
pillow==10.4.0
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
python-docx==1.1.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
tldextract==5.1.2
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
